fix: common entity
